Removed finalize() methods from all Java objects.
Finalize is a method that's called by the Java Finalizer thread, which runs
concurrently with everything else in the JVM. Therefore finalize() methods, if
implemented, are re-entrant. However RAMCloud code, in general, is not thread
safe, and problems can occur if clean-up is attempted while other RAMCloud
objects which are related are being used or in a way that's out of order with
expected execution. An example is if finalize is called on a RAMCloudTransaction
object while the RAMCloud object is being used, or after the RAMCloud object has
already been deleted.