Increased ensureServers timeout to 60s.
Sometimes when starting a cluster of 8 or more masters, ensureServers
would timeout waiting for a master or two. Looking at the logs, it
appears that there isn't anything wrong with those masters, but they do
take about 30 seconds to setup and pin all their memory (50G out of 60G
total in these experiments). This means that as the number of servers
increases, the probability that one of them will take longer than 30
seconds to complete setup increases. Seems like at 8 servers, 1 or 2
tend to not make it.