Changed read only tx to not register on server.
Read only transactions no longer register with the TransactionManager in
MasterService::txPrepare(). This was leading to unnecessary transaction timeouts
because read only transactions are not committed by the client after the prepare
phase. While this did not affect correctness, it was resulting in unnecessary
work for the server, especially if a client executes a long stream of read-only
transactions.